agent/requirements.txt
Flask
geopy
langchain
python-dotenv
SQLAlchemy
firebase_admin
googlemaps
langchain-google-genai
google-generativeai
